LibRef.html #1327 HTML fixes and #1340 corrections
Added description of List Deconstruction.
Say that unicode() takes any expression.
Expand back the example for parseAsInt(), which must have lost the final part.
Expand explanation of maxBy and minBy, now that max and min are gone.
A big block of ten lines:
\<el-code>\<el-type>Float\</el-code>\</el-type>
which should be:
\<el-code>\<el-type>Float\</el-type>\</el-code>
Plus a few typos.