whack-a-mole.html change wording re time interval and fix HTML
I noticed that the heading says "Move the mole randomly, once a second"
but the timeout is now, wisely, set to two seconds.
People may find this confusing so I added an extra sentence:

"This can be reduced later to say 1000 (one second) to make the game more challenging once you have got the hang of it and finished testing."

And changed some later text to tie in with the fact that the timeout is already two seconds.
Feel free to change it again if necessary.