LangRef.html #1340 fill in "Procedure call" section
I wrote some text, with examples, to go in the "Procedure call" section.
Feel free to correct and clarify it if necessary.
I'm not quite sure of the rules for the code markup, so I hope
it is good enough.
And I fixed one HTML tag mismatch elsewhere, which I had to fix so that
my tag checker gave the all clear, and it did find a missing \</p> tag
in the text that I had added!