LangRef.html LibRef.html Concepts.html (new) #1340 Random, getKey, keyword ids etc
Quite a lot changed but I decided to make it all one commit, as easier to manage, assuming that it doesn't all need unravelling.
But it is in three parts.
Rather long, but I like to have a record of what I changed.

1. LangRef.html #1340 add some more keywords as id's

I erroneously changed an "if" and an "else" link from href="#if" to href="#if_statement" to make it work in a previous commit.  Now I realise that a better route is to add id="if" and id="else" in extra little div's as we need them for links from the IDE.
So I have edited those two lines back how they were (#else and #if) before that change (Commit ae65b05)
(not sure if GitHub intelligently links commit IDs: https://github.com/elan-language/LanguageAndIDE/commit/ae65b052ac5c03dca91c549f8e5312cd174e107a).
I have added quite a lot of extra `<div id="whatever"></div>` id's for keywords for the IDE to use as targets.
I'm not sure that it will use all of them, but no harm if a few extra.
A few still to consider, some of which need some text writing: image, out, ref, this (I have put id "this"  on the "class" section, but there is no description of "this").  I don't know how we handle links to LibRef eg for "image".
PS I have put a section about "ref" in LibRef, but I expect the IDE will refer to LangRef.html#ref if anywhere so that will need sorting out.

Also I broke id "identifier" in one place.  Add <div id="identifier"> so that "identifier" and "Identifer" both work.  We can always tidy it up in slow time.

2. LibRef.html #1340 Type Random, getKey etc, function ref, ids

- Add section on the type Random, ported from the old Elan 6 manual
- Add section "Reading keys ‘on the fly’", ported from the old Elan 6 manual
- Remove duplicate table entries for getKey and pressAnyKeyToContinue
- Add extra ids for Stack_functions and Queue_functions so it works and maintains the convention in the row where the hrefs appear. It can be tidied once we decide on a convention.
- Add extra id for `<div id="tail"></div>` on <h3 class="no-TOC" id="ListDeconstruction">List Deconstruction</h3>.  There is a reference in ElanIndex.html.
- Fill in the section "Passing a function as a reference", copied from the old Elan 6 manual. This is actually a language feature not a library feature, but there was a heading already there in LibRef.html, where higher-order functions are discussed.  Good enough for the moment.

3. New file documentation/Concepts.html #1340

Bernard suggested that we need another document which covers topics which don't really live in LangRef.html or LibRef.html, so I have created Concepts.html.
Partly on the principle that it sometimes helps to write *something* to clarify the options.
Three of the sections I wrote (or ported across from the Elan 6 manual) ended up being more suitable for LangRef or LibRef so I dropped them in there.
That makes the Concepts document rather thin.
Also I wasn't sure what was wanted, so I didn't write too much in case I was on the wrong track.
I am happy to have it published -- I don't think there is anything libellous in there.
And other people may suggest topics.