Range and double index type checking #1942
Indices in range (eg a[3..6]) and double (Array2D eg a[2, 3]) were not being type-checked properly,
eg a["2".."6"] was being accepted.
I had to change a few class properties from private to public; I hope that is OK.
Two tests added.
The change that I suggested in part three of the Issue caused a lot of test failures
so I have backed it out (src/compiler/compile-rules.ts).